+++ CHANGELOG.md
@@ -7,6 +7,8 @@
         * `MicroUtils`: `0.4.11` -> `0.4.15`
         * `Klock`: `2.0.1` -> `2.0.2`
         * `Ktor`: `1.4.3` -> `1.5.0`
+* `Core`:
+    * All bot actions got functions for short calling, like `recordVideo` for `RecordVideoNote`
 
 ## 0.30.10